Defining a specific String case on source of Charge.create
I'm really not sure what I'm running into here. Passing in `source` as a
String was still triggering the when branch to call id on it. Even
though it's not any of those types.

Adding a String speicific case keeps the flow moving forward as
expected. But to even make it more interesting, the order matters.
Placing it after the Object expections still doesn't fix the problem. It
has to be first to make the compiler happy.

It's a hacky fix but it does work...